# perf(codec): optimize encode/decode functions for modern CPUs