--- a/.github/workflows/actionlint.yml
+++ b/.github/workflows/actionlint.yml
@@ -42,7 +42,7 @@ jobs:
     name: zizmor 🌈
     runs-on: ubuntu-24.04
     permissions:
-      security-events: write
+      security-events: write # Needs to upload SARIF info.
     steps:
       - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8 # v5.0.0
         with: