--- a/.github/workflows/bencher-upload.yml
+++ b/.github/workflows/bencher-upload.yml
@@ -15,8 +15,8 @@ jobs:
     # The bench workflow fails when a regression is detected, but we want to upload results regardless.
     if: github.event.workflow_run.conclusion == 'success' || github.event.workflow_run.conclusion == 'failure'
     permissions:
-      pull-requests: write
-      checks: write
+      pull-requests: write # Writes comments on PRs.
+      checks: write # Writes comments on commits.
     runs-on: ubuntu-24.04
     env:
       BENCHER_PROJECT: ${{ github.event.repository.name }}