--- a/.github/workflows/pr-comment.yml
+++ b/.github/workflows/pr-comment.yml
@@ -19,7 +19,7 @@ jobs:
   comment:
     name: Comment on PR
     permissions:
-      pull-requests: write
+      pull-requests: write # Writes comments on PRs.
     runs-on: ubuntu-24.04
     if: |
       github.event.workflow_run.event == 'pull_request' &&