--- a/.github/workflows/qns.yml
+++ b/.github/workflows/qns.yml
@@ -31,7 +31,7 @@ jobs:
     outputs:
       imageID: ${{ steps.docker_build_and_push.outputs.imageID }}
     permissions:
-      packages: write
+      packages: write # To push to GitHub Container Registry.
     steps:
       - uses: docker/setup-qemu-action@29109295f81e9208d7d86ff1c6c12d2833863392 # v3.6.0
         with: