@larseggert 2025-09-17
So Claude hallucinated checking the assembly with godbolt.org :-) Second attempt with more explicit instructions.