@larseggert 2025-09-17
Is this worth it?