@larseggert 2025-09-17
> prepend a commit
Sorry, not sure what you mean here?