@mxinden 2025-09-18
Sorry for the brevity. Let me expand.

- `cargo mutant` is able to alter the code in significant ways without test failures.
- Thus I assume this code isn't well tested.
- Given that this pull request does substantial changes to the logic, it might introduce a bug, e.g. in the edge cases.
- Thus I suggest first writing a commit increasing test coverage, making sure it succeeds on `main`.
- Then running the test on this patch to make sure it still succeeds.