@larseggert 2025-09-18
So I think the `mutants` check here is meaningless, because for the code in question
```rust
            () if v < (1 << 6) => self.encode_uint(1, v),
            () if v < (1 << 14) => self.encode_uint(2, v | (1 << 14)),
            () if v < (1 << 30) => self.encode_uint(4, v | (2 << 30)),
            () if v < (1 << 62) => self.encode_uint(8, v | (3 << 62)),
```
or (`|`) and xor (`^`) produce the same result. We're taking a value smaller than some boundary, and or-ing or xor-ing it with a value larger than the boundary. Same result.